docs(readme): re-organize README and write about my decisions
The new organization has two main sections: Usage and Development. By the moment Usage section is empty because there is not any useful to be used.

In Development section there are information about my decisions related to technologies or how to used them, references and there will be instructions about scripts, tools or how to collaborate.